requirements-gpu.txt
asttokens
atomicwrites
attrs
cdflib
certifi
colorama
cupy-cuda110
executing
fastrlock
h5py
importlib-metadata
littleutils
llvmlite
more-itertools
numba
numpy
nvidia-ml-py3
packaging
pluggy
py
pynvml
pyparsing
pytest
scipy
six
sorcery
tifffile
wcwidth
wincertstore
wrapt
zipp
napari[all]
magicgui
pyqtgraph
colorcet
matplotlib